Change IoE Controller to SuperAdmin Controller for GetModeratorsByIoEId and fix mistakes.
