+++ YIF.Core.Domain/ApiModels/ResponseApiModels/IoEModeratorsResponseApiModel.cs
@@ -3,15 +3,6 @@
     public class IoEModeratorsResponseApiModel
     {
         /// <summary>
-        /// Unique Id
-        /// </summary>
-        ///
-        public string Id { get; set; }
-        /// <summary>
-        /// Admin Id
-        /// </summary>
-        public string AdminId { get; set; }
-        /// <summary>
         /// User Id
         /// </summary>
         public string UserId { get; set; }
