+++ YIF.Core.Domain/Repositories/InstitutionOfEducationModeratorRepository.cs
@@ -1,5 +1,4 @@
 ﻿using AutoMapper;
-using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -7,7 +6,6 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
 using YIF.Core.Data.Entities;
-using YIF.Core.Data.Entities.IdentityEntities;
 using YIF.Core.Data.Interfaces;
 using YIF.Core.Domain.DtoModels.EntityDTO;
 
@@ -17,14 +15,11 @@ namespace YIF.Core.Domain.Repositories
     {
         private readonly IApplicationDbContext _dbContext;
         private readonly IMapper _mapper;
-        private readonly UserManager<DbUser> _userManager;
         public InstitutionOfEducationModeratorRepository(
             IMapper mapper,
-            UserManager<DbUser> userManager,
             IApplicationDbContext dbContext)
         {
             _mapper = mapper;
-            _userManager = userManager;
             _dbContext = dbContext;
         }
         public async Task<string> AddUniModerator(InstitutionOfEducationModerator institutionOfEducationModerator)
@@ -44,7 +39,7 @@ namespace YIF.Core.Domain.Repositories
              _dbContext.Dispose();
         }
 
-        public async Task<IEnumerable<InstitutionOfEducationModeratorDTO>> GetAllUniModerators()
+        public async Task<IEnumerable<InstitutionOfEducationModeratorDTO>> GetAll()
         {
             var moderators = await _dbContext.InstitutionOfEducationModerators
                 .Include(x => x.User)
@@ -65,22 +60,12 @@ namespace YIF.Core.Domain.Repositories
             return _mapper.Map<IEnumerable<InstitutionOfEducationModeratorDTO>>(result);
         }
 
-        public Task<InstitutionOfEducationModeratorDTO> GetById(string id)
-        {
-            throw new NotImplementedException();
-        }
-
         public Task<bool> Update(InstitutionOfEducationModerator item)
         {
             throw new NotImplementedException();
         }
 
         public Task<InstitutionOfEducationModeratorDTO> Get(string id)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<IEnumerable<InstitutionOfEducationModeratorDTO>> GetAll()
         {
             throw new NotImplementedException();
         }
