+++ YIF.Core.Domain/ServiceInterfaces/IInstitutionOfEducationService.cs
@@ -16,7 +16,5 @@ namespace YIF.Core.Domain.ServiceInterfaces
         Task<IEnumerable<string>> GetInstitutionOfEducationAbbreviations(FilterApiModel filterModel);
         Task AddInstitutionOfEducationToFavorite(string institutionOfEducationId, string userId);
         Task DeleteInstitutionOfEducationFromFavorite(string institutionOfEducationId, string userId);
-        Task<ResponseApiModel<IEnumerable<IoEModeratorsResponseApiModel>>> GetIoEModeratorsByIoEId(string ioEId);
-
     }
 }
