+++ YIF.Core.Service/Concrete/Services/SuperAdminService.cs
@@ -38,6 +38,7 @@ namespace YIF.Core.Service.Concrete.Services
         private readonly ISpecialtyRepository<Specialty, SpecialtyDTO> _specialtyRepository;
         private readonly ISchoolAdminRepository<SchoolAdminDTO> _schoolAdminRepository;
         private readonly ISchoolModeratorRepository<SchoolModeratorDTO> _schoolModeratorRepository;
+        private readonly IInstitutionOfEducationModeratorRepository<InstitutionOfEducationModerator, InstitutionOfEducationModeratorDTO> _ioEModeratorRepository;
         private readonly ITokenRepository<TokenDTO> _tokenRepository;
         private readonly ResourceManager _resourceManager;
         private readonly IWebHostEnvironment _env;
@@ -57,6 +58,7 @@ namespace YIF.Core.Service.Concrete.Services
             ISpecialtyRepository<Specialty, SpecialtyDTO> specialtyRepository,
             ISchoolAdminRepository<SchoolAdminDTO> schoolAdminRepository,
             ISchoolModeratorRepository<SchoolModeratorDTO> schoolModeratorRepository,
+            IInstitutionOfEducationModeratorRepository<InstitutionOfEducationModerator, InstitutionOfEducationModeratorDTO> ioEModeratorRepository,
             ITokenRepository<TokenDTO> tokenRepository,
             ResourceManager resourceManager, 
             IWebHostEnvironment env,
@@ -75,6 +77,7 @@ namespace YIF.Core.Service.Concrete.Services
             _specialtyRepository = specialtyRepository;
             _schoolAdminRepository = schoolAdminRepository;
             _schoolModeratorRepository = schoolModeratorRepository;
+            _ioEModeratorRepository = ioEModeratorRepository;
             _tokenRepository = tokenRepository;
             _resourceManager = resourceManager;
             _env = env;
@@ -297,7 +300,6 @@ namespace YIF.Core.Service.Concrete.Services
             }
         }
 
-        ///<inheritdoc/>
         public async Task<PageResponseApiModel<InstitutionOfEducationAdminResponseApiModel>> GetAllInstitutionOfEducationAdmins(
             InstitutionOfEducationAdminSortingModel institutionOfEducationAdminFilterModel,
             PageApiModel pageModel)
@@ -318,6 +320,7 @@ namespace YIF.Core.Service.Concrete.Services
 
             return result;
         }
+
         public async Task<ResponseApiModel<IEnumerable<SchoolAdminResponseApiModel>>> GetAllSchoolAdmins()
         {
             var result = new ResponseApiModel<IEnumerable<SchoolAdminResponseApiModel>>();
@@ -330,11 +333,6 @@ namespace YIF.Core.Service.Concrete.Services
             return result.Set(true);
         }
 
-        /// <summary>
-        /// Used only for InstitutionOfEducationAdminDTO
-        /// </summary>
-        /// <param name="institutionOfEducationAdminFilterModel"></param>
-        /// <returns>Ordered list</returns>
         private async Task<IEnumerable<InstitutionOfEducationAdminDTO>> GetAllInstitutionOfEducationAdminsSorted(
             InstitutionOfEducationAdminSortingModel institutionOfEducationAdminFilterModel)
         {
@@ -386,6 +384,15 @@ namespace YIF.Core.Service.Concrete.Services
 
             return result.Set(new DescriptionResponseApiModel(_resourceManager.GetString("SpecialtyWasSuccessefullyChanged")),
                 await _specialtyRepository.Update(_mapper.Map<Specialty>(specialtyDTO)));
+        }
+
+        public async Task<ResponseApiModel<IEnumerable<IoEModeratorsResponseApiModel>>> GetIoEModeratorsByIoEId(string ioEId)
+        {
+            return new ResponseApiModel<IEnumerable<IoEModeratorsResponseApiModel>>
+            {
+                Object = _mapper.Map<IEnumerable<IoEModeratorsResponseApiModel>>(await _ioEModeratorRepository.GetByIoEId(ioEId)),
+                Success = true
+            };
         }
     }
 }
