+++ YIF_XUnitTests/Integration/YIF_Backend/Controllers/InstitutionOfEducationControllerTests.cs
@@ -2,7 +2,6 @@
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json.Linq;
-using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
 using YIF_XUnitTests.Integration.Fixture;
@@ -180,19 +179,6 @@ namespace YIF_XUnitTests.Integration.YIF_Backend.Controllers
             Assert.Equal("application/json; charset=utf-8",
                 response.Content.Headers.ContentType.ToString());
             Assert.True(contentJsonObj.Count >= 1);
-        }
-
-        [Fact]
-        public async void GetModeratorsByIoEId_EndpointReturnsListOfModeratorsWithOkStatusCode_IfEverythingIsOk()
-        {
-            // Arrange
-            var ioEId = _context.InstitutionOfEducations.FirstOrDefault().Id;
-
-            // Act
-            var response = await _client.GetAsync($"api/InstitutionOfEducation/GetModeratorsById?ioEId={ioEId}");
-
-            // Assert
-            response.EnsureSuccessStatusCode();
         }
         #endregion
